Re-organize the code

+++ Windows Build Identifier/Identification/InstalledImage/DetectionHandler.cs
@@ -26,101 +26,13 @@ using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Linq;
+using WindowsBuildIdentifier.Interfaces;
 
-namespace WindowsBuildIdentifier
+namespace WindowsBuildIdentifier.Identification.InstalledImage
 {
-    public class IdentifyWindowsInstall
+    public class DetectionHandler
     {
-        public enum BuildType
-        {
-            fre,
-            chk
-        }
-
-        public enum Type
-        {
-            Client,
-            Server,
-            ServerV
-        }
-
-        public enum Licensing
-        {
-            Retail,
-            OEM,
-            Volume
-        }
-
-        public class Report
-        {
-            public ulong MajorVersion;
-            public ulong MinorVersion;
-            public ulong BuildNumber;
-            public ulong DeltaVersion;
-            public string BranchName;
-            public string CompileDate;
-            public string Tag;
-            public MachineType Architecture;
-            public BuildType BuildType;
-            public HashSet<Type> Type = new HashSet<Type>();
-            public string Sku;
-            public string[] Editions;
-            public Licensing Licensing;
-            public string[] LanguageCodes;
-        }
-
-        public enum MachineType : ushort
-        {
-            unknown = 0x0,
-            am33 = 0x1d3,
-            amd64 = 0x8664,
-            arm = 0x1c0,
-            arm64 = 0xaa64,
-            woa = 0x1c4,
-            ebc = 0xebc,
-            x86 = 0x14c,
-            ia64 = 0x200,
-            m32r = 0x9041,
-            mips16 = 0x266,
-            mipsfpu = 0x366,
-            mipsfpu16 = 0x466,
-            powerpc = 0x1f0,
-            powerpcfp = 0x1f1,
-            r4000 = 0x166,
-            sh3 = 0x1a2,
-            sh3dsp = 0x1a3,
-            sh4 = 0x1a6,
-            sh5 = 0x1a8,
-            thumb = 0x1c2,
-            wcemipsv2 = 0x169,
-        }
-
-        private class VersionInfo1
-        {
-            public ulong MajorVersion;
-            public ulong MinorVersion;
-            public ulong BuildNumber;
-            public ulong DeltaVersion;
-            public string BranchName;
-            public string CompileDate;
-            public MachineType Architecture;
-            public BuildType BuildType;
-        }
-
-        private class VersionInfo2
-        {
-            public ulong MajorVersion;
-            public ulong MinorVersion;
-            public ulong BuildNumber;
-            public ulong DeltaVersion;
-            public string BranchName;
-            public string CompileDate;
-            public string Tag;
-            public Licensing Licensing;
-            public string[] LanguageCodes;
-        }
-
-        public static Report IdentifyWindows(WindowsInstallProviderInterface installProvider)
+        public static Report IdentifyWindowsNT(WindowsInstallProviderInterface installProvider)
         {
             Report report = new Report();
 
@@ -162,13 +74,18 @@ namespace WindowsBuildIdentifier
             bool IsUnstaged = !fileentries.Any(x => x.Contains(@"system32\ntoskrnl.exe", StringComparison.InvariantCultureIgnoreCase)) &&
                 !fileentries.Any(x => x.Contains(@"system32\ntkrnlmp.exe", StringComparison.InvariantCultureIgnoreCase));
 
-            Console.WriteLine("Extracting version information from the image 1");
-            VersionInfo1 info = ExtractVersionInfo(kernelPath);
+            VersionInfo1 info = new VersionInfo1();
 
-            File.Delete(kernelPath);
+            if (!string.IsNullOrEmpty(kernelPath))
+            {
+                Console.WriteLine("Extracting version information from the image 1");
+                info = ExtractVersionInfo(kernelPath);
 
-            report.Architecture = info.Architecture;
-            report.BuildType = info.BuildType;
+                File.Delete(kernelPath);
+
+                report.Architecture = info.Architecture;
+                report.BuildType = info.BuildType;
+            }
 
             if (IsUnstaged)
             {
@@ -215,7 +132,7 @@ namespace WindowsBuildIdentifier
                 CompileDate = info2.CompileDate
             };
 
-            WindowsVersion correctVersion = GetGreaterVersion(version1, version2);
+            WindowsVersion correctVersion = Common.GetGreaterVersion(version1, version2);
 
             report.MajorVersion = correctVersion.MajorVersion;
             report.MinorVersion = correctVersion.MinorVersion;
@@ -307,111 +224,26 @@ namespace WindowsBuildIdentifier
             return report;
         }
 
-        private static MachineType GetMachineTypeFromFile(Stream fs)
-        {
-            BinaryReader br = new BinaryReader(fs);
-            fs.Seek(0x3c, SeekOrigin.Begin);
-            Int32 peOffset = br.ReadInt32();
-            fs.Seek(peOffset, SeekOrigin.Begin);
-            UInt32 peHead = br.ReadUInt32();
-            if (peHead != 0x00004550) // "PE\0\0", little-endian
-            {
-                throw new Exception("Can't find PE header");
-            }
-
-            MachineType machineType = (MachineType)br.ReadUInt16();
-            br.Close();
-            fs.Close();
-            return machineType;
-        }
-
-        private static VersionInfo1 ParseBuildString(VersionInfo1 verinfo, string BuildString)
-        {
-            if (BuildString.Contains(" built by: ") && BuildString.Contains(" at: ") && BuildString.Count(x => x == '.') == 3)
-            {
-                // Early version
-                var splitparts = BuildString.Split(' ');
-                var rawversion = splitparts[0];
-                var branch = splitparts[3];
-                var compiledate = splitparts[5];
-
-                var splitver = rawversion.Split('.');
-
-                verinfo.MajorVersion = ulong.Parse(splitver[0]);
-                verinfo.MinorVersion = ulong.Parse(splitver[1]);
-                verinfo.BuildNumber = ulong.Parse(splitver[2]);
-                verinfo.DeltaVersion = ulong.Parse(splitver[3]);
-
-                verinfo.BranchName = branch;
-                verinfo.CompileDate = compiledate;
-            }
-            else if (BuildString.Contains(" (") && BuildString.Contains(")") && BuildString.Count(x => x == '.') >= 4)
-            {
-                if (BuildString.Contains(" (WinBuild."))
-                {
-                    // MS new thing
-
-                    var splitparts = BuildString.Split(' ');
-                    var rawversion = splitparts[0];
-
-                    var splitver = rawversion.Split('.');
-
-                    verinfo.MajorVersion = ulong.Parse(splitver[0]);
-                    verinfo.MinorVersion = ulong.Parse(splitver[1]);
-                    verinfo.BuildNumber = ulong.Parse(splitver[2]);
-                    verinfo.DeltaVersion = ulong.Parse(splitver[3]);
-                }
-                else
-                {
-                    // Normal thing
-
-                    var splitparts = BuildString.Split(' ');
-                    var rawversion = splitparts[0];
-                    var rawdetails = splitparts[1];
-
-                    var splitdetails = rawdetails.TrimStart('(').TrimEnd(')').Split('.');
-
-                    var branch = splitdetails[0];
-                    var compiledate = splitdetails[1];
-
-                    var splitver = rawversion.Split('.');
-
-                    verinfo.MajorVersion = ulong.Parse(splitver[0]);
-                    verinfo.MinorVersion = ulong.Parse(splitver[1]);
-                    verinfo.BuildNumber = ulong.Parse(splitver[2]);
-                    verinfo.DeltaVersion = ulong.Parse(splitver[3]);
-
-                    verinfo.BranchName = branch;
-                    verinfo.CompileDate = compiledate;
-                }
-            }
-
-
-            return verinfo;
-        }
-
         private static VersionInfo1 ExtractVersionInfo(string kernelPath)
         {
             VersionInfo1 result = new VersionInfo1();
 
             FileVersionInfo info = FileVersionInfo.GetVersionInfo(kernelPath);
 
-            result.Architecture = GetMachineTypeFromFile(new FileStream(kernelPath, FileMode.Open));
+            result.Architecture = Common.GetMachineTypeFromFile(new FileStream(kernelPath, FileMode.Open));
 
-            result = ParseBuildString(result, info.FileVersion);
+            WindowsVersion verinfo = Common.ParseBuildString(info.FileVersion);
+
+            result.MajorVersion = verinfo.MajorVersion;
+            result.MinorVersion = verinfo.MinorVersion;
+            result.BuildNumber = verinfo.BuildNumber;
+            result.DeltaVersion = verinfo.DeltaVersion;
+            result.BranchName = verinfo.BranchName;
+            result.CompileDate = verinfo.CompileDate;
+
             result.BuildType = info.IsDebug ? BuildType.chk : BuildType.fre;
 
             return result;
-        }
-
-        private class WindowsVersion
-        {
-            public ulong MajorVersion;
-            public ulong MinorVersion;
-            public ulong BuildNumber;
-            public ulong DeltaVersion;
-            public string BranchName;
-            public string CompileDate;
         }
 
         private static VersionInfo2 ExtractVersionInfo2(string softwareHivePath, string systemHivePath)
@@ -541,72 +373,6 @@ namespace WindowsBuildIdentifier
             }
 
             return result;
-        }
-
-        private static WindowsVersion GetGreaterVersion(WindowsVersion version1, WindowsVersion version2)
-        {
-            if (version1.MajorVersion != version2.MajorVersion)
-            {
-                if (version1.MajorVersion > version2.MajorVersion)
-                {
-                    return version1;
-                }
-                return version2;
-            }
-
-            if (version1.MinorVersion != version2.MinorVersion)
-            {
-                if (version1.MinorVersion > version2.MinorVersion)
-                {
-                    return version1;
-                }
-                return version2;
-            }
-
-            if (version1.BuildNumber != version2.BuildNumber)
-            {
-                if (version1.BuildNumber > version2.BuildNumber)
-                {
-                    return version1;
-                }
-                return version2;
-            }
-
-            if (version1.DeltaVersion != version2.DeltaVersion)
-            {
-                if (version1.DeltaVersion > version2.DeltaVersion)
-                {
-                    return version1;
-                }
-                return version2;
-            }
-
-            if (version1.CompileDate != version2.CompileDate)
-            {
-                if (string.IsNullOrEmpty(version1.CompileDate))
-                {
-                    return version2;
-                }
-                if (string.IsNullOrEmpty(version2.CompileDate))
-                {
-                    return version1;
-                }
-
-                CultureInfo provider = CultureInfo.InvariantCulture;
-
-                string format = "yyMMDD-HHmm";
-
-                DateTime date1 = DateTime.ParseExact(version1.CompileDate, format, provider);
-                DateTime date2 = DateTime.ParseExact(version2.CompileDate, format, provider);
-
-                if (date1 > date2)
-                {
-                    return version1;
-                }
-                return version2;
-            }
-
-            return version1;
         }
 
         private static string ExtractEditionInfo(string systemHivePath)
@@ -753,38 +519,5 @@ namespace WindowsBuildIdentifier
 
             return report.ToArray();
         }
-
-        public static void DisplayReport(Report report)
-        {
-            string typedisp = report.Type != null
-                ? string.Join(", ", report.Type.Select(e => e.ToString()))
-                : "";
-
-            string editiondisp = report.Editions != null
-                ? string.Join(", ", report.Editions.Select(e => e.ToString()))
-                : "";
-
-            string langdisp = report.LanguageCodes != null
-                ? string.Join(", ", report.LanguageCodes.Select(e => e.ToString()))
-                : "";
-
-            Console.WriteLine();
-            Console.WriteLine("MajorVersion : " + report.MajorVersion);
-            Console.WriteLine("MinorVersion : " + report.MinorVersion);
-            Console.WriteLine("BuildNumber  : " + report.BuildNumber);
-            Console.WriteLine("DeltaVersion : " + report.DeltaVersion);
-            Console.WriteLine("BranchName   : " + report.BranchName);
-            Console.WriteLine("CompileDate  : " + report.CompileDate);
-            Console.WriteLine("Tag          : " + report.Tag);
-            Console.WriteLine("Architecture : " + report.Architecture);
-            Console.WriteLine("BuildType    : " + report.BuildType);
-            Console.WriteLine("Type         : " + typedisp);
-            Console.WriteLine("Sku          : " + report.Sku);
-            Console.WriteLine("Editions     : " + editiondisp);
-            Console.WriteLine("Licensing    : " + report.Licensing);
-            Console.WriteLine("LanguageCodes: " + langdisp);
-            Console.WriteLine();
-        }
-
     }
 }
